Source Code/requirements.txt
matplotlib==3.8.2
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
scikit_learn==1.3.2
scipy==1.13.1
seaborn==0.13.2
skimage==0.0
